requirements.txt
Django==5.0.2
gTTS==2.5.1
python-docx==1.1.0
pydub==0.25.1